+++ test/datanorm/file_test.rb
@@ -15,6 +15,9 @@ class TheFileTest < Minitest::Test
     file = Datanorm::File.new(path: TestAsset.v4_with_texts)
 
     assert_equal 4, file.header.version.number
+    assert_equal 'Artikelstammdaten Deutschland, 38_DE EXAMPLE EXAMPLETEC GMBH ' \
+                 'www.example.de info@example.de',
+                 file.header.title
   end
 
   def test_v4_lines
@@ -25,11 +28,21 @@ class TheFileTest < Minitest::Test
     assert_predicate lines[0], :kind_product?
   end
 
+  def test_v5_lines
+    file = Datanorm::File.new(path: TestAsset.v5_first_texts_then_products)
+
+    lines = file.to_a
+
+    assert_predicate lines[0], :kind_text?
+    assert_equal 'Example GmbH Produktdaten', file.header.title
+  end
+
   def test_v4_invalid_lines
     file = Datanorm::File.new(path: TestAsset.v4_with_empty_lines_and_invalid_tags)
 
     products = file.to_a
 
     assert_equal 6, products.size
+    assert_equal 'Preisliste examples GmbH', file.header.title
   end
 end