+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.0.4
+
+- Parse header title
+
+# 1.0.3
+
+- Improve missing reference detection
+
 # 1.0.2
 
 - Ignore empty lines in Datanorm files instead of causing an exception.